Overwrite modulemap, apinotes files
Before creating the symlinks to the modulemap and the apinotes files, delete the ones that might be previously there to avoid mklink from failing. Since we are just dealing with CI machines, I haven't checked that we are not deleting some user-generated files, but even outside CI, that would be weird.